+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Changelog for massaction
 
 ## Unreleased
+- NEW : Add mass action to create a new supplier proposal from a customer proposal or order - *09/10/2025* - 1.7.0
 
 ## Release 1.6
 - FIX : Compat V22 - *02/10/2025* - 1.6.4